jlowin-fastmcp/pyproject.toml
[project]
name = "fastmcp"
version = "0.1.0"
description = "Self-Evolving Digital Twin MCP Server"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "pydantic>=1.8.0",
    "python-dotenv>=0.19.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "langchain>=0.1.0",
    "tiktoken>=0.5.0",
    "numpy>=1.21.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "sentence-transformers>=2.2.0",
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=fastmcp"

[tool.coverage.run]
source = ["src/fastmcp"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "pass",
    "@abstractmethod",
]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "B", "I"] 